Trying to improve thread safety, getting bugs while playing

+++ src/main/java/org/gp3/Playable.java
@@ -7,5 +7,5 @@ public interface Playable {
     String getAlbum();
     String getGenre();
     String getYear();
-    long getDurationMillis();
+    double getDurationMillis();
 }
